Fixed a bug where player online where throwin a exception because int is too small

+++ Api/InternalTypes/PlayerData.cs
@@ -178,7 +178,7 @@ namespace DestinyPCL.InternalTypes
         public object raceHash { get; set; }
         public object genderHash { get; set; }
         public object classHash { get; set; }
-        public int currentActivityHash { get; set; }
+        public long currentActivityHash { get; set; }
         public int lastCompletedStoryHash { get; set; }
         public Stats stats { get; set; }
         public Customization customization { get; set; }
